Add files via upload
Draft 1 in Github, draft 3 in Jupyter NBs
I am not sure if I need more of this for the project? I am interested in asking more questions about the data(please see 'Capstone02_pseudo_code'). But I don't know if that is the best use of my time right now.