Add files via upload
Pseudo code for this project. This file encountered an error when running a simple method (dict), and then was unable to save. I copied what I needed over (see 'drafts'). This is the pseudo code for other questions and interests I have regarding the data. I wonder how much of it is worth asking. I want to come back to this later and keep working on it, even start data analysis.